[hipDNN] Add descale (block scale dequantize) CPU reference implementation
Add CPU reference for the block scale dequantize operation so that
test_sdk can validate GPU engine outputs against a known-correct CPU
implementation.

New files:
- CpuFpReferenceBlockScaleDequantize.hpp: core math Y[i] = X[i] * scale[block_of(i)]
  with support for negative scale mode (Y = X * 2^(-scale))
- BlockScaleDequantizePlan.hpp: plan executor and builder bridging flatbuffer
  graph nodes to the CPU reference
- BlockScaleDequantizeSignatureKey.hpp: registry key with 9 type combinations
  including FP8 E4M3/E5M2 with E8M0 scale support

Modified files:
- FlatbufferDatatypeMapping: add FP8_E4M3, FP8_E5M2, FP8_E8M0 type mappings
- CpuFpReferenceUtilities: add FP8 types to IS_VALID_TENSOR_TYPE_V and safeConvert
- Tensor.hpp: add FP8_E8M0 to createTensor() factory
- CpuFpReferenceValidation: fix FP8 subtraction by casting to float first
- PlanRegistrySignatureKey/PlanBuilderRegistry/CpuReferenceGraphExecutor: wire
  BlockScaleDequantize into the existing registry infrastructure

Tests: 15 new tests covering CPU reference math, plan execution, and signature keys.

+++ projects/hipdnn/test_sdk/include/hipdnn_test_sdk/utilities/CpuFpReferenceValidation.hpp
@@ -51,7 +51,7 @@ public:
             T refValue = refView.getHostValue(indices);
             T implValue = implView.getHostValue(indices);
 
-            auto absDiff = static_cast<float>(fabs(implValue - refValue));
+            auto absDiff = fabs(static_cast<float>(implValue) - static_cast<float>(refValue));
             auto threshold
                 = _absoluteTolerance + _relativeTolerance * fabs(static_cast<float>(refValue));
